requirements.txt
Django>=5,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
psycopg2-binary
djoser>=2.2
